setup simple DAW animation
