+++ frontend/package.json
@@ -10,6 +10,7 @@
     "react-dom": "^18.3.1",
     "react-scripts": "5.0.1",
     "recorder-js": "^1.0.7",
+    "tone": "^15.0.4",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
