+++ frontend/src/App.js
@@ -1,148 +1,46 @@
-import React, { useState} from "react";
-
-//need to pick a better click sound
-const clickSound = new Audio("https://actions.google.com/sounds/v1/alarms/beep_short.ogg");
+import React, { useState } from "react";
+import DAWGrid from "./components/DAWGrid";
+import PlaybackControls from "./components/PlaybackControls";
+import Recorder from "./components/Recorder";
 
 function App() {
   const [bpm, setBpm] = useState(120); // Default BPM
-  const [isRecording, setIsRecording] = useState(false);
-  const [isCountingDown, setIsCountingDown] = useState(false);
-  const [audioStream, setAudioStream] = useState(null); 
-  const [audioDownloadLink, setAudioDownloadLink] = useState(null);
-  
-  const handleBpmChange = (event) => {
-    setBpm(event.target.value);
-  };
-  const getBeatDuration = () => {
-    return (60 / bpm) * 1000; // ms per beat
+  const [isPlaying, setIsPlaying] = useState(false);
+  const [isMetronomeOn, setIsMetronomeOn] = useState(false); // State for metronome toggle
+
+  const play = () => {
+    setIsPlaying(true);
   };
 
-  //needs to be awaited to play the sound
-  const playClick = async () => {
-    clickSound.currentTime = 0;
-    await clickSound.play();
+  const pause = () => {
+    setIsPlaying(false);
   };
 
-
-  const startCountdown = () => {
-    setIsCountingDown(true);
-    const beatDuration = getBeatDuration();
-    let beatCount = 0;
-
-    const clickInterval = setInterval(async () => {
-      if (beatCount >= 4) {
-        clearInterval(clickInterval);
-        startRecording();
-      }
-      else{
-
-      await playClick();  // Play the click sound
-      beatCount += 1;
-    }
-
-    }, beatDuration);  
+  // Toggle the metronome on or off
+  const toggleMetronome = () => {
+    setIsMetronomeOn(!isMetronomeOn);
   };
-
-
-
-const startRecording = async () => {
-  setIsCountingDown(false);
-  setIsRecording(true);
-  const stream = await navigator.mediaDevices.getUserMedia({ audio: true });
-  setAudioStream(stream); 
-
-  const recorder = new MediaRecorder(stream, { mimeType: "audio/webm" });
-  const chunks = [];
-  
-  recorder.onstop = () => {
-    console.log("Recording stopped. Total chunks captured:", chunks.length);
-    setIsRecording(false);
-    if (audioStream) {
-      audioStream.getTracks().forEach(track => track.stop()); 
-    }
-
-    const blob = new Blob(chunks, { type: recorder.mimeType });
-    const audioURL = window.URL.createObjectURL(blob);
-    setAudioDownloadLink(audioURL);
-  };
-
-  recorder.ondataavailable = (e) => {
-    chunks.push(e.data);
-  };
-
-  recorder.start();
-
-  //stops recording after 4 beats
-  const beatDuration = getBeatDuration();
-  setTimeout(() => {
-    recorder.stop();
-  }, 4 * beatDuration);
-
-};
-
-
-const saveRecording = async () => {
-  if (audioDownloadLink) {
-    const blob = await fetch(audioDownloadLink).then(res => res.blob());  // Fetch the audio Blob
-
-    const formData = new FormData();
-    formData.append('file', blob, 'recording.webm');  // Append the file as 'file'
-
-    // Send the form data to the backend
-    try {
-      const response = await fetch('http://127.0.0.1:8000/upload-audio/', {
-        method: 'POST',
-        body: formData,
-      });
-      
-      if (response.ok) {
-        const data = await response.json();
-        console.log('File uploaded successfully:', data.file_url);
-      } else {
-        console.error('File upload failed.');
-      }
-    } catch (error) {
-      console.error('Error uploading file:', error);
-    }
-  } else {
-    console.error("No recording available to save.");
-  }
-};
 
   return (
     <div className="App">
-      <h1>BPM: {bpm}</h1>
+      <h1>DAW Interface</h1>
 
-      {/* BPM Slider */}
-      <div>
-        <label>Set BPM: {bpm}</label>
-        <input
-          type="range"
-          min="30"
-          max="240"
-          value={bpm}
-          onChange={handleBpmChange}
-        />
-      </div>
+      {/* DAW Grid with Playhead and BPM */}
+      <DAWGrid isPlaying={isPlaying} bpm={bpm} isMetronomeOn={isMetronomeOn} />
 
-      {/* Recording Button */}
-      <div>
-        <button
-          onClick={startCountdown}
-          disabled={isRecording || isCountingDown}
-          style={{
-            backgroundColor: isRecording ? "red" : "gray",
-            color: "white",
-          }}
-        >
-          {isRecording ? "Recording..." : isCountingDown ? "Get Ready..." : "Start Recording"}
-        </button>
-      </div>
+      {/* Playback Controls */}
+      <PlaybackControls
+        bpm={bpm}
+        setBpm={setBpm}
+        isPlaying={isPlaying}
+        play={play}
+        pause={pause}
+        isMetronomeOn={isMetronomeOn} // Pass metronome state
+        toggleMetronome={toggleMetronome} // Pass toggle function
+      />
 
-      {/* Save Recording Button */}
-      {!isRecording && (
-        <button onClick={saveRecording}>Save Recording</button>
-      )}
+      {/* Recorder Component */}
+      <Recorder bpm={bpm} />
     </div>
   );
 }
