+++ frontend/src/components/DAWGrid.css
@@ -1,0 +1,24 @@
+.daw-grid {
+    display: grid;
+    grid-template-rows: repeat(8, 1fr); /* 8 rows */
+    gap: 4px; /* Adjust gap between cells if needed */
+    margin: 20px auto;
+    max-width: 640px;
+  }
+  
+  .daw-row {
+    display: grid;
+    grid-template-columns: repeat(16, 1fr); /* 8 columns */
+    gap: 4px; /* Adjust gap between cells if needed */
+  }
+  
+  .daw-cell {
+    width: 50px;  /* Increased square size */
+    height: 50px; /* Increased square size */
+    background-color: lightgray;
+    border: 1px solid #ddd;
+  }
+  
+  .playhead {
+    background-color: red;
+  }
